Move Events to exist as a resource
