+++ packages/thyseus/src/events/EventRegistryKey.ts
@@ -1,0 +1,1 @@
+export const EventRegistryKey = Symbol('EventRegistry');
