+++ packages/thyseus/src/world/World.ts
@@ -2,8 +2,7 @@ import { bits, DEV_ASSERT, Memory } from '../utils';
 import { WorldBuilder, type Registry } from './WorldBuilder';
 import { Commands } from '../commands';
 import { Entities, Table } from '../storage';
-import { EventReader, EventWriter } from '../events';
-import { ComponentRegistryKey, EventRegistryKey } from './registryKeys';
+import { ComponentRegistryKey } from './registryKeys';
 import { StartSchedule, type ExecutorInstance } from '../schedule';
 import {
 	validateAndCompleteConfig,
@@ -39,8 +38,6 @@ export class World {
 	#archetypeToTable: Map<bigint, Table> = new Map<bigint, Table>();
 	queries: Query<any, any>[] = [];
 	resources: object[] = [];
-	eventReaders: EventReader<any>[] = [];
-	eventWriters: EventWriter<any>[] = [];
 
 	schedules: Record<symbol, ExecutorInstance> = {};
 
@@ -76,35 +73,6 @@ export class World {
 
 		this.entities = new Entities(this);
 		this.commands = new Commands(this);
-
-		const eventTypes = registry.get(EventRegistryKey) as
-			| Set<Struct>
-			| undefined;
-		if (eventTypes) {
-			let eventsPointer = this.threads.queue(() =>
-				Memory.alloc(EventReader.size * eventTypes.size),
-			);
-			for (const eventType of eventTypes) {
-				const id = this.eventReaders.length;
-				this.eventReaders.push(
-					new EventReader(
-						this.commands,
-						eventType,
-						eventsPointer,
-						id,
-					),
-				);
-				this.eventWriters.push(
-					new EventWriter(
-						this.commands,
-						eventType,
-						eventsPointer,
-						id,
-					),
-				);
-				eventsPointer += EventReader.size;
-			}
-		}
 	}
 
 	get isMainThread(): boolean {
