+++ packages/thyseus/src/world/registryKeys.ts
@@ -1,3 +1,2 @@
 export const ComponentRegistryKey = Symbol('ComponentRegistry');
-export const EventRegistryKey = Symbol('EventRegistry');
 export const ResourceRegistryKey = Symbol('ResourceRegistry');
